# OTSS_User_Service/openapi.yaml
openapi: 3.1.0
info:
  title: OTSS User Service
  version: 0.2.0
  description: >
    Manages user registration via WSO2 SCIM2, IAM-synced user profiles,
    and support-staff availability data used by Queue/Ticket services.
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /staff:
    get:
      summary: Discover staff for assignment
      tags: [Staff]
      parameters:
        - in: query
          name: category
          schema: { $ref: '#/components/schemas/TicketCategory' }
        - in: query
          name: isOnline
          schema: { type: boolean }
        - in: query
          name: hasCapacity
          schema: { type: boolean, description: Return only staff with load < maxLoad }
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Matching staff status rows with linked profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StaffStatusWithProfile'
  /staff/isonline/{profileId}:
    patch:
      summary: Update staff availability and capacity
      tags: [Staff]
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStaffStatusDto'
      responses:
        '200':
          description: Updated staff status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffStatusWithProfile'
        '404':
          description: Staff profile missing or not marked as staff
...
components:
...
  schemas:
    UserRole:
      type: string
      enum: [END_USER, STAFF, ADMIN]
    TicketCategory:
      type: string
      enum: [IT, BILLING, ACCESS, OTHER]
...
    StaffProfileInputDto:
      type: object
      properties:
        isOnline: { type: boolean }
        currentLoad:
          type: integer
          minimum: 0
        maxLoad:
          type: integer
          minimum: 1
          maximum: 50
          default: 5
        categories:
          type: array
          items: { $ref: '#/components/schemas/TicketCategory' }
      additionalProperties: false
    UpdateStaffStatusDto:
      $ref: '#/components/schemas/StaffProfileInputDto'
    Profile:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        phoneNumber: { type: string }
        timezone: { type: string }
        language: { type: string }
        role: { $ref: '#/components/schemas/UserRole' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        staffStatus:
          $ref: '#/components/schemas/StaffStatus'
      additionalProperties: false
    StaffStatus:
      type: object
      properties:
        id: { type: string, format: uuid }
        profileId: { type: string }
        isOnline: { type: boolean }
        currentLoad: { type: integer }
        maxLoad: { type: integer }
        categories:
          type: array
          items: { $ref: '#/components/schemas/TicketCategory' }
        updatedAt: { type: string, format: date-time }
      additionalProperties: false
    StaffStatusWithProfile:
      type: object
      properties:
        id: { type: string, format: uuid }
        profileId: { type: string }
        isOnline: { type: boolean }
        currentLoad: { type: integer }
        maxLoad: { type: integer }
        categories:
          type: array
          items: { $ref: '#/components/schemas/TicketCategory' }
        updatedAt: { type: string, format: date-time }
        profile:
          $ref: '#/components/schemas/Profile'
      additionalProperties: false
...